refactor: split crypto UI owner
