+++ service-worker.js
@@ -348,6 +348,7 @@ const APP_SHELL = [
   '/js/recommendations.js',
   '/js/recommendations-region.js',
   '/js/crypto.js',
+  '/js/crypto-ui.js',
   '/js/data-wipe.js',
   '/js/backup-cycle.js',
   '/js/backup-serialization.js',
